requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.0
djangorestframework==3.9.1
httpie==1.0.2
idna==2.8
mysqlclient==1.4.2
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1

